#173 added tooltip ID variable to QC autoplots & Content.Rmd
Allow user to adjust Content.Rmd to control which ID shown in the tooltip. Default passed is 'SampleLabelColumn' from the results object. the new variable is defaulted to SampleName which frequently is the RCC file ID in case this call is removed from the Content.Rmd. Boxplots in autoplot have not been updated to use tooltipID